Avoid mutating shared project configuration.
The current implementation mutates the host project's configuration.
For older versions of ember-cli the project configuration was generated
every time `project.config()` was invoked, but newer versions will avoid
the wasted/duplicated work and begin caching. Unfortunately, ember-cli-eyeglass
relies on the fact that `project.config()` is not cached and mutates the
`project.config().eyeglass` config directly.

This commit ensures that the project config is cloned before we begin
mutating.